fix:删除规格类型做校验，被使用的不能删除 fix：销售整车列表接口传参修改

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/service/specificat/impl/SpecificatTypeServiceImpl.java
@@ -10,10 +10,12 @@ import com.redescooter.ses.tool.utils.DateUtil;
 import com.redescooter.ses.tool.utils.SesStringUtils;
 import com.redescooter.ses.web.ros.constant.SequenceName;
 import com.redescooter.ses.web.ros.dao.specificat.SpecificatTypeServiceMapper;
+import com.redescooter.ses.web.ros.dm.OpeSaleScooter;
 import com.redescooter.ses.web.ros.dm.OpeSpecificatDef;
 import com.redescooter.ses.web.ros.dm.OpeSpecificatType;
 import com.redescooter.ses.web.ros.exception.ExceptionCodeEnums;
 import com.redescooter.ses.web.ros.exception.SesWebRosException;
+import com.redescooter.ses.web.ros.service.base.OpeSaleScooterService;
 import com.redescooter.ses.web.ros.service.base.OpeSpecificatTypeService;
 import com.redescooter.ses.web.ros.service.specificat.SpecificatDefService;
 import com.redescooter.ses.web.ros.service.specificat.SpecificatTypeService;
@@ -53,6 +55,10 @@ public class SpecificatTypeServiceImpl implements SpecificatTypeService {
 
     @Autowired
     private SpecificatTypeServiceMapper specificatTypeServiceMapper;
+
+
+    @Autowired
+    private OpeSaleScooterService opeSaleScooterService;
 
 
     @Override
@@ -203,6 +209,13 @@ public class SpecificatTypeServiceImpl implements SpecificatTypeService {
     @Transactional
     public GeneralResult specificatTypeDelete(IdEnter enter) {
         // todo 校验当前规格是否被使用
+        // 2020 10 16 追加规格类型是否被销售车辆使用了，使用不能删除
+        QueryWrapper<OpeSaleScooter> qw = new QueryWrapper<>();
+        qw.eq(OpeSaleScooter.COL_SPECIFICAT_ID,enter.getId());
+        int count = opeSaleScooterService.count(qw);
+        if(count > 0){
+            throw new SesWebRosException(ExceptionCodeEnums.SPECIFICAT_TYPE_IS_USED.getCode(), ExceptionCodeEnums.SPECIFICAT_TYPE_IS_USED.getMessage());
+        }
         opeSpecificatTypeService.removeById(enter.getId());
         // 删除规格的自定义项
         specificatDefService.deleSpecificatDef(enter.getId());
